Split FutureExt into two traits
One half of the original trait is only relevant for `Tasks`, so it can
be private to daemon.

The other half is used in `daemon-tests`, but to reduce the API
footprint of `daemon`, we inline the functionality in the one place
where it is used in `daemon-tests`.
