+++ daemon/src/lib.rs
@@ -10,7 +10,6 @@ use crate::model::OpeningFee;
 use crate::model::Price;
 use crate::model::Usd;
 use crate::oracle::Attestation;
-use crate::tokio_ext::FutureExt;
 use address_map::Stopping;
 use anyhow::Context;
 use anyhow::Result;
@@ -28,6 +27,7 @@ use std::net::SocketAddr;
 use std::time::Duration;
 use time::ext::NumericalDuration;
 use tokio::sync::watch;
+use tokio_ext::FutureExt;
 use xtra::message_channel::StrongMessageChannel;
 use xtra::Actor;
 use xtra::Address;
@@ -48,6 +48,7 @@ pub mod command;
 pub mod connection;
 pub mod db;
 pub mod fan_out;
+mod future_ext;
 pub mod keypair;
 pub mod maker_cfd;
 pub mod maker_inc_connections;
